Attempt to move exec_tags to the `benchmark<...>` class.
The goal was to move the `cpu_only` functionality into exec_tags rather
than using a runtime variable in `benchmark_base`.

Currently, we cannot just specify an `nvbench::exec_tag::cpu_only` in
the KernelGenerator, because we need to know if the benchmark is
CPU-only when generating states, assigning default devices, etc. since
we only want to run CPU-only benchmarks a single time, not once per
detected device. Since `exec_tag`s are passed into `state.exec` in the
current design, there's no way for the benchmark to know whether or not
the benchmark is CPU-only early enough. So, the we use a runtime flag.

The issue with trying to move `exec_tagg` into `benchmark<...>` is that
`state.exec` needs to combine two pieces of compile-time information:

1. The `exec_tag` type for determining which measurement classes to
    instantiate.

2. The `KernelLauncher` type so that we can inline the `KernelLauncher`
    code while instantiating the measurement classes.

`nvbench::state` only has a pointer to the `benchmark_base` virtual
class, so we'd need to either (a) retrieve the `exec_tag` information
via a virtual call on `benchmark_base` or (b) pass the
`KernelLauncher` as a function pointer through a virtual function on
`benchmark_base`.

The former would require doing a dynamic dispatch on the exec_tag, which
defeats the purpose of having static `exec_tag` information in the first
place. The latter would prevent inlining of the `KernelLauncher`.

Neither of these are great solutions.

I do like the `NVBENCH_CREATE` design in this PR. Might come back for
that at some point.